fix: re-apply anchor on anchor
