+++ src/__tests__/Stylist/utils.test.js
@@ -1,4 +1,4 @@
-import { remove, merge, when } from '../../Stylist/utils'
+import { when } from '../../Stylist/utils'
 
 test('when', () => {
   const cond1 = () => true
@@ -15,26 +15,4 @@ test('when', () => {
     .run()
 
   expect(output).toBe(1)
-})
-
-test('remove item from array', () => {
-  const array = [1, 2, 3]
-  const result = remove(array, 2)
-  expect(result).not.toBe(array)
-  expect(result).toEqual([1, 3])
-})
-
-test('merge an element to an array', () => {
-  const array = [1, 2]
-  let result = merge(array, 3)
-  expect(result).not.toBe(array)
-  expect(result).toEqual([1, 2, 3])
-
-  result = merge(result, 3)
-  expect(result).toEqual([1, 2, 3])
-})
-
-test('merge element with no array', () => {
-  const result = merge(undefined, 3)
-  expect(result).toEqual([3])
 })
